Refactor more exports
