+++ sites/docs/src/pages/_JsonHighlight.js
@@ -4,10 +4,9 @@ import { useColorMode } from '@docusaurus/theme-common';
 import Highlight, { defaultProps } from 'prism-react-renderer';
 import copy from 'copy-text-to-clipboard';
 import { getHighlightTheme } from './_highlight-theme';
+import { JSON_TRANSLATION_KEY } from './_editor-utils';
 
 import styles from './styles.module.css';
-
-export const JSON_TRANSLATION_KEY = 'vitessceJsonTranslation';
 
 export default function JsonHighlight(props) {
   const { json } = props;
