+++ sites/docs/src/pages/_editor-utils.js
@@ -1,0 +1,1 @@
+export const JSON_TRANSLATION_KEY = 'vitessceJsonTranslation';
